Move ssh private_key, public_key, username to Server model to reduce duplication
