+++ lib/fog/aws/models/compute/spot_request.rb
@@ -1,10 +1,10 @@
-require 'fog/core/model'
+require 'fog/compute/models/server'
 
 module Fog
   module Compute
     class AWS
       
-      class SpotRequest < Fog::Model
+      class SpotRequest < Fog::Compute::Server
 
         identity :id,                          :aliases => 'spotInstanceRequestId'
 
@@ -33,10 +33,6 @@ module Fog
         attribute :fault,                      :squash  => 'message'
         attribute :user_data
         
-        attr_writer   :private_key, :private_key_path, :public_key, :public_key_path
-
-        attr_writer :username
-
         def initialize(attributes={})
           self.groups ||= ["default"]
           self.flavor_id ||= 't1.micro'
@@ -73,24 +69,6 @@ module Fog
 
         def key_pair=(new_keypair)
           self.key_name = new_keypair && new_keypair.name
-        end
-        
-        def private_key_path
-          @private_key_path ||= Fog.credentials[:private_key_path]
-          @private_key_path &&= File.expand_path(@private_key_path)
-        end
-
-        def private_key
-          @private_key ||= private_key_path && File.read(private_key_path)
-        end
-
-        def public_key_path
-          @public_key_path ||= Fog.credentials[:public_key_path]
-          @public_key_path &&= File.expand_path(@public_key_path)
-        end
-
-        def public_key
-          @public_key ||= public_key_path && File.read(public_key_path)
         end
 
         def ready?
